refactoring and made app test and got snapshot
