+++ src/components/Calculator.js
@@ -1,5 +1,6 @@
 import React, { useState } from 'react';
 import calculate from '../logic/calculate';
+import Navbar from './Navbar';
 import '../BasicStyle/Calculator.css';
 
 const Calculator = () => {
@@ -31,47 +32,49 @@ const Calculator = () => {
   };
 
   return (
-    <main className="calculator-wrapper">
-      <div className="title">
-        <h1>Let&apos;s do some math!</h1>
-      </div>
-      <section className="calculator-container">
-        <div
-          id="display"
-          className="display"
-        >
-          { display }
+    <main>
+      <Navbar />
+      <div className="calculator-wrapper">
+        <div className="title">
+          <h1>Let&apos;s do some math!</h1>
         </div>
-
-        <div className="numbers-container">
-          <button
-            className="clear-display"
-            type="button"
-            id="clear"
-            onClick={calculation}
+        <section className="calculator-container">
+          <div
+            id="display"
+            className="display"
           >
-            AC
-          </button>
+            { display }
+          </div>
 
-          <button
-            className="plus-minus"
-            id="pm"
-            type="button"
-            onClick={calculation}
-          >
-            +/-
-          </button>
+          <div className="numbers-container">
+            <button
+              className="clear-display"
+              type="button"
+              id="clear"
+              onClick={calculation}
+            >
+              AC
+            </button>
 
-          <button
-            className="percent"
-            id="percents"
-            type="button"
-            onClick={calculation}
-          >
-            %
-          </button>
+            <button
+              className="plus-minus"
+              id="pm"
+              type="button"
+              onClick={calculation}
+            >
+              +/-
+            </button>
 
-          {
+            <button
+              className="percent"
+              id="percents"
+              type="button"
+              onClick={calculation}
+            >
+              %
+            </button>
+
+            {
             numbers.map((number) => (
               <button
                 type="button"
@@ -85,18 +88,18 @@ const Calculator = () => {
             ))
         }
 
-          <button
-            className="decimals"
-            id="decimal"
-            type="button"
-            onClick={calculation}
-          >
-            .
-          </button>
-        </div>
+            <button
+              className="decimals"
+              id="decimal"
+              type="button"
+              onClick={calculation}
+            >
+              .
+            </button>
+          </div>
 
-        <div className="operators-container">
-          {
+          <div className="operators-container">
+            {
             operators.map((operator) => (
               <button
                 className="orange operators"
@@ -109,8 +112,9 @@ const Calculator = () => {
               </button>
             ))
         }
-        </div>
-      </section>
+          </div>
+        </section>
+      </div>
     </main>
   );
 };
